+++ src/lib/api/estimate/requests/createDesignatedEstimate.ts
@@ -1,7 +1,7 @@
 import { tokenFetch } from "@/lib/utils/fetch-client";
 
 export async function createDesignatedEstimate(moverId: string, requestId: string) {
-  return await tokenFetch(`/movers/${moverId}`, {
+  return await tokenFetch(`/requests/movers/${moverId}`, {
     method: 'PATCH',
     body: JSON.stringify({ requestId }),
     headers: {
